requirements.txt
dm-haiku
einops
ftfy
google-cloud-storage
numpy
optax
regex
Pillow
torch
torchvision
tqdm
